+++ marketapi/build.gradle
@@ -33,8 +33,9 @@ dependencies {
 	// mariadb
 	implementation group: 'org.mariadb.jdbc', name: 'mariadb-java-client', version: '2.4.1'
 
-	// QueryDsl : 스프링 부트 3.0 이상
+	// queryDsl 관련 코드생성 기능 제공
 	implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
+	// queryDsl 라이브러리
 	annotationProcessor "com.querydsl:querydsl-apt:5.0.0:jakarta"
 	annotationProcessor "jakarta.annotation:jakarta.annotation-api"
 	annotationProcessor "jakarta.persistence:jakarta.persistence-api"
@@ -67,21 +68,6 @@ dependencies {
 
 tasks.named('test') {
 	useJUnitPlatform()
-}
-
-// === ⭐ QueryDsl 빌드 옵션 (선택) ===
-def querydslDir = "$buildDir/generated/querydsl"
-
-sourceSets {
-	main.java.srcDirs += [ querydslDir ]
-}
-
-tasks.withType(JavaCompile) {
-	options.annotationProcessorGeneratedSourcesDirectory = file(querydslDir)
-}
-
-clean.doLast {
-	file(querydslDir).deleteDir()
 }
 
 jar {
