+++ marketapi/src/main/java/com/example/marketapi/auth/domain/PrincipalDetails.java
@@ -1,0 +1,45 @@
+package com.example.marketapi.auth.domain;
+
+import com.example.marketapi.member.domain.Member;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+import java.util.Collection;
+import java.util.Collections;
+import java.util.Map;
+
+public record PrincipalDetails(Member member,
+                               Map<String, Object>attributes,
+                               String attributeKey) implements OAuth2User, UserDetails {
+
+    @Override
+    public String getName() {
+        return null;
+    }
+
+    @Override
+    public Map<String, Object> getAttributes() {
+        return null;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return Collections.singletonList(
+                new SimpleGrantedAuthority(member.getRole().getKey())
+        );
+    }
+
+    @Override
+    public String getPassword() {
+        return null;
+    }
+
+    @Override
+    public String getUsername() {
+        return null;
+    }
+
+
+}
