+++ marketapi/src/main/java/com/example/marketapi/auth/dto/LoginResponse.java
@@ -1,0 +1,6 @@
+package com.example.marketapi.auth.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record LoginResponse (@NotBlank String accessToken){
+}
