+++ marketapi/src/main/java/com/example/marketapi/auth/jwt/TokenAuthenticationFilter.java
@@ -1,0 +1,58 @@
+package com.example.marketapi.auth.jwt;
+
+import static org.springframework.http.HttpHeaders.AUTHORIZATION;
+
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Component;
+import org.springframework.util.ObjectUtils;
+import org.springframework.util.StringUtils;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import java.io.IOException;
+
+@Component
+@RequiredArgsConstructor
+public class TokenAuthenticationFilter extends OncePerRequestFilter {
+
+    private final TokenProvider tokenProvider;
+    public static final String TOKEN_PREFIX = "Bearer ";
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+        String accessToken = resolveToken(request);
+
+        if (tokenProvider.validateToken(accessToken)){
+            setAuthentication(accessToken);
+        }
+        else{
+            String reissueAccessToken = tokenProvider.reissueAccessToken(accessToken);
+
+            if (StringUtils.hasText(reissueAccessToken)) {
+                setAuthentication(reissueAccessToken);
+
+                response.setHeader(AUTHORIZATION, TOKEN_PREFIX + reissueAccessToken);
+            }
+        }
+
+        filterChain.doFilter(request, response);
+    }
+
+    private void setAuthentication(String accessToken) {
+        Authentication authentication = tokenProvider.getAuthentication(accessToken);
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+    }
+
+    private String resolveToken(HttpServletRequest request) {
+        String token = request.getHeader(AUTHORIZATION);
+        if (ObjectUtils.isEmpty(token) || !token.startsWith(TOKEN_PREFIX)){
+            return null;
+        }
+        return token.substring(TOKEN_PREFIX.length());
+    }
+}
