+++ marketapi/src/main/java/com/example/marketapi/auth/jwt/TokenExceptionFilter.java
@@ -1,0 +1,22 @@
+package com.example.marketapi.auth.jwt;
+
+
+import com.example.marketapi.auth.jwt.exception.TokenException;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import java.io.IOException;
+
+public class TokenExceptionFilter extends OncePerRequestFilter {
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+        try {
+            filterChain.doFilter(request, response);
+        } catch (TokenException e) {
+            response.sendError(e.getErrorCode().getStatus().value(), e.getMessage());
+        }
+    }
+}
