+++ marketapi/src/main/java/com/example/marketapi/auth/jwt/exception/TokenException.java
@@ -1,0 +1,10 @@
+package com.example.marketapi.auth.jwt.exception;
+
+import com.example.marketapi.global.exception.CustomException;
+import com.example.marketapi.global.exception.ErrorCode;
+
+public class TokenException extends CustomException {
+    public TokenException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
