+++ marketapi/src/main/java/com/example/marketapi/auth/service/CustomOAuth2UserService.java
@@ -1,0 +1,47 @@
+package com.example.marketapi.auth.service;
+
+import com.example.marketapi.member.domain.Member;
+import com.example.marketapi.member.repository.MemberRepository;
+import com.example.marketapi.auth.domain.PrincipalDetails;
+import com.example.marketapi.auth.dto.OAuth2UserInfo;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
+import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
+import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Map;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class CustomOAuth2UserService extends DefaultOAuth2UserService {
+
+    private final MemberRepository memberRepository;
+
+    @Transactional
+    @Override
+    public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException{
+        Map<String, Object> oauth2UserAttributes = super.loadUser(userRequest).getAttributes();
+
+        String registrationId = userRequest.getClientRegistration().getRegistrationId();
+
+        String usernameAttributeName = userRequest.getClientRegistration().getProviderDetails()
+                .getUserInfoEndpoint().getUserNameAttributeName();
+
+        OAuth2UserInfo oAuth2UserInfo = OAuth2UserInfo.of(registrationId, oauth2UserAttributes);
+
+        Member member = getOrSave(oAuth2UserInfo);
+
+        return new PrincipalDetails(member, oauth2UserAttributes, usernameAttributeName);
+    }
+
+    private Member getOrSave(OAuth2UserInfo oAuth2UserInfo) {
+        Member member = memberRepository.findByEmail(oAuth2UserInfo.email())
+                .orElseGet(oAuth2UserInfo::toEntity);
+        return memberRepository.save(member);
+    }
+}
