--- marketapi/src/main/java/com/example/marketapi/jwt/JwtAccessDeniedHandler.java
@@ -1,19 +1,0 @@
-package com.example.marketapi.jwt;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-public class JwtAccessDeniedHandler implements AccessDeniedHandler {
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
-        // 403
-        response.sendError(HttpServletResponse.SC_FORBIDDEN);
-    }
-}
