--- marketapi/src/main/java/com/example/marketapi/jwt/JwtAuthenticationEntryPoint.java
@@ -1,19 +1,0 @@
-package com.example.marketapi.jwt;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
-        // 401
-        response.sendError(HttpServletResponse.SC_UNAUTHORIZED);
-    }
-}
