--- marketapi/src/main/java/com/example/marketapi/jwt/TokenExceptionFilter.java
@@ -1,22 +1,0 @@
-package com.example.marketapi.jwt;
-
-
-import com.example.marketapi.jwt.exception.TokenException;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-public class TokenExceptionFilter extends OncePerRequestFilter {
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-        try {
-            filterChain.doFilter(request, response);
-        } catch (TokenException e) {
-            response.sendError(e.getErrorCode().getStatus().value(), e.getMessage());
-        }
-    }
-}
