--- marketapi/src/main/java/com/example/marketapi/jwt/exception/TokenException.java
@@ -1,10 +1,0 @@
-package com.example.marketapi.jwt.exception;
-
-import com.example.marketapi.global.exception.CustomException;
-import com.example.marketapi.global.exception.ErrorCode;
-
-public class TokenException extends CustomException {
-    public TokenException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
