--- marketapi/src/main/java/com/example/marketapi/jwt/repository/TokenRepository.java
@@ -1,12 +1,0 @@
-package com.example.marketapi.jwt.repository;
-
-import com.example.marketapi.jwt.entity.Token;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface TokenRepository extends CrudRepository<Token, String> {
-    Optional<Token> findByAccessToken(String accessToken);
-}
