--- marketapi/src/main/java/com/example/marketapi/jwt/service/TokenService.java
@@ -1,40 +1,0 @@
-package com.example.marketapi.jwt.service;
-
-import com.example.marketapi.global.exception.CustomException;
-import com.example.marketapi.global.exception.ErrorCode;
-import com.example.marketapi.jwt.entity.Token;
-import com.example.marketapi.jwt.repository.TokenRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-public class TokenService {
-
-    private final TokenRepository tokenRepository;
-
-    public void deleteRefreshToken(String memberKey){
-        tokenRepository.deleteById(memberKey);
-    }
-
-    @Transactional
-    public void saveOrUpdate(String memberKey, String refreshToken, String accessToken){
-        Token token = tokenRepository.findByAccessToken(accessToken)
-                .map(o -> o.updateRefreshToken(refreshToken))
-                .orElseGet(() -> new Token(memberKey, refreshToken, accessToken));
-
-        tokenRepository.save(token);
-    }
-
-    public Token findByAccessTokenOrThrow(String accessToken) {
-        return tokenRepository.findByAccessToken(accessToken)
-                .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND, "TOKEN_EXPIRED"));
-    }
-
-    @Transactional
-    public void updateToken(String accessToken, Token token) {
-        token.updateAccessToken(accessToken);
-        tokenRepository.save(token);
-    }
-}
