--- marketapi/src/main/java/com/example/marketapi/login/controller/AuthController.java
@@ -1,39 +1,0 @@
-package com.example.marketapi.login.controller;
-
-import com.example.marketapi.jwt.JwtFilter;
-import com.example.marketapi.jwt.TokenProvider;
-import com.example.marketapi.jwt.service.TokenService;
-import com.example.marketapi.login.dto.LoginDto;
-import com.example.marketapi.login.dto.LoginResponse;
-import com.example.marketapi.login.dto.TokenDto;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/api/auth")
-@RequiredArgsConstructor
-public class AuthController {
-
-    private final TokenService tokenService;
-
-    @GetMapping("/success")
-    public ResponseEntity<LoginResponse> loginSuccess(@Valid LoginResponse loginresponse){
-        return ResponseEntity.ok(loginresponse);
-    }
-
-    @DeleteMapping("/logout")
-    public ResponseEntity<Void> logout(@AuthenticationPrincipal UserDetails userDetails) {
-        tokenService.deleteRefreshToken(userDetails.getUsername());
-        return ResponseEntity.noContent().build();
-    }
-}
