--- marketapi/src/main/java/com/example/marketapi/login/dto/LoginResponse.java
@@ -1,6 +1,0 @@
-package com.example.marketapi.login.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record LoginResponse (@NotBlank String accessToken){
-}
