--- marketapi/src/main/java/com/example/marketapi/login/service/RedisMessageService.java
@@ -1,11 +1,0 @@
-package com.example.marketapi.login.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class RedisMessageService {
-    private final RedisTemplate<String, Object> redisTemplate;
-}
