--- marketapi/src/main/java/com/example/marketapi/oauth2/CustomOAuth2UserService.java
@@ -1,54 +1,0 @@
-package com.example.marketapi.oauth2;
-
-import com.example.marketapi.login.repository.UserRepository;
-import com.example.marketapi.member.domain.Member;
-import com.example.marketapi.member.repository.MemberRepository;
-import com.example.marketapi.oauth2.dto.OAuth2UserInfo;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.AuthorityUtils;
-import org.springframework.security.oauth2.client.registration.ClientRegistration;
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
-import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.LinkedHashMap;
-import java.util.LinkedHashSet;
-import java.util.List;
-import java.util.Map;
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class CustomOAuth2UserService extends DefaultOAuth2UserService {
-
-    private final MemberRepository memberRepository;
-
-    @Transactional
-    @Override
-    public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException{
-        Map<String, Object> oauth2UserAttributes = super.loadUser(userRequest).getAttributes();
-
-        String registrationId = userRequest.getClientRegistration().getRegistrationId();
-
-        String usernameAttributeName = userRequest.getClientRegistration().getProviderDetails()
-                .getUserInfoEndpoint().getUserNameAttributeName();
-
-        OAuth2UserInfo oAuth2UserInfo = OAuth2UserInfo.of(registrationId, oauth2UserAttributes);
-
-        Member member = getOrSave(oAuth2UserInfo);
-
-        return new PrincipalDetails(member, oauth2UserAttributes, usernameAttributeName);
-    }
-
-    private Member getOrSave(OAuth2UserInfo oAuth2UserInfo) {
-        Member member = memberRepository.findByEmail(oAuth2UserInfo.email())
-                .orElseGet(oAuth2UserInfo::toEntity);
-        return memberRepository.save(member);
-    }
-}
